+++ src/tools/lib/plot.h
@@ -15,7 +15,8 @@ OQMC_CABI bool oqmc_plot_zoneplate(const char* sampler, int nsamples,
 
 // NOLINTNEXTLINE: C style naming
 OQMC_CABI bool oqmc_plot_error(const char* shape, const char* sampler,
-                               int nsequences, int nsamples, float* out);
+                               int dimensionA, int dimensionB, int nsequences,
+                               int nsamples, float* out);
 
 // NOLINTNEXTLINE: C style naming
 OQMC_CABI bool oqmc_plot_error_filter_space(const char* shape,